search.php fix
Search page wasn’t working. I copied all the files down from their
server and all files were the same as the repo except the functions.php
file was “newer” so it did an overwrite. Which was fine. Then I changed
the code in the search.php to the code I used on another project in my
search.php. it was different and seems to work better.